+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # native-agreement Changelog
 
+## 1.0.1 (2021-04-26)
+
+#### Bug Fix
+
+- improved scroll precision
+
 ## 1.0.0 (2021-04-07)
 
 #### New Feature
